rpc/jsonrpc: adjust log index for bor transaction (#52)
On devnets, the log index of bor transaction receipts turned out to be incorrect.

We used to fetch the last transaction index and get log index and cumulative gas used against that index. On debugging, it turns out txNum+1 was used which used to return log index including the state-sync logs. Later, we used to exclude the total logs in state-sync receipt to get to the first log index of that transaction. This led to incorrect results in almost all scenarios where normal and state-sync transactions were present in a block (irrespective of the hard fork).

We now instead use txNum to get the first log index which can be used in the next receipt. This simplifies the calculation and results into correct index in logs of bor receipts. The old logic is still kept for future reference and debugging.